refactor(settings): separate aria viewmodel commands
